refactor: share InflightSource family between FlattenConcat and FlattenMerge
Motivation:
After the previous commit, FlattenMerge grew its own copy of the
`InflightSource[T]` hierarchy (Iterator/Range/Repeat/CompletedFuture/
PendingFuture) duplicating what FlattenConcat already had in its
companion object. Two near-identical families across two files is a
maintenance hazard: any future tweak to the value-presented optimization
(e.g. adding a new source type, fixing a Java-stream cleanup leak) would
have to be mirrored, and the families had already drifted in small ways
(e.g. `tryPull`/`cancel`/`materialize` declared abstract in concat with
no-op overrides on every subclass; concat used `isClosed = true` for the
completed-future variant while merge used `!_hasNext`).

Modification:
- Extract the common `InflightSource[T]` base and the five value-presented
  subclasses (Iterator/Range/Repeat/CompletedFuture/PendingFuture) into
  a new `pekko.stream.impl.fusing.InflightSources` package-private object.
- Promote `tryPull` / `cancel` / `materialize` from abstract to concrete
  no-op defaults, so the value-presented subclasses no longer carry empty
  overrides. Stages that wrap a real `SubSinkInlet` (only FlattenConcat's
  `attachAndMaterializeSource` does this) override what they need.
- Align `InflightCompletedFutureSource.isClosed` to FlattenConcat's
  `true` semantics — behaviorally equivalent in both stages, but more
  faithful to the source being a one-shot cached value.
- Drop the `sealed` modifier on `InflightSource` so FlattenConcat's
  attached-substream anonymous subclass can still extend it from another
  file in the same package.
- Remove the duplicate definitions from FlattenConcat's companion (now
  unused, drop the empty companion entirely) and from FlattenMerge's
  companion. Both stages import from the shared object instead.

Result:
- Net -176 lines of duplication; one canonical home for the
  optimization's data types.
- Future additions (e.g. extending the optimization to other stream-of-
  streams stages such as `MergeMany`-style operators) only need to
  reference `InflightSources`.
- All FlattenConcat / FlattenMerge / flatMapConcat parallelism tests
  remain green; MiMa is clean (`@InternalApi private[fusing]`).
