+++ stream/src/main/scala/org/apache/pekko/stream/impl/fusing/FlattenConcat.scala
@@ -27,102 +27,10 @@ import pekko.stream.{ Attributes, FlowShape, Graph, Inlet, Outlet, SourceShape, 
 import pekko.stream.impl.{ Buffer => BufferImpl, FailedSource, JavaStreamSource, TraversalBuilder }
 import pekko.stream.impl.Stages.DefaultAttributes
 import pekko.stream.impl.fusing.GraphStages.{ FutureSource, RepeatSource, SingleSource }
+import pekko.stream.impl.fusing.InflightSources._
 import pekko.stream.scaladsl.Source
 import pekko.stream.stage.{ GraphStage, GraphStageLogic, InHandler, OutHandler }
 import pekko.util.OptionVal
-
-/**
- * INTERNAL API
- */
-@InternalApi
-private[pekko] object FlattenConcat {
-  private sealed abstract class InflightSource[T] {
-    def hasNext: Boolean
-    def next(): T
-    def tryPull(): Unit
-    def cancel(cause: Throwable): Unit
-    def isClosed: Boolean
-    def hasFailed: Boolean = failure.isDefined
-    def failure: Option[Throwable] = None
-    def materialize(): Unit = ()
-  }
-
-  private final class InflightIteratorSource[T](iterator: Iterator[T]) extends InflightSource[T] {
-    override def hasNext: Boolean = iterator.hasNext
-    override def next(): T = iterator.next()
-    override def tryPull(): Unit = ()
-    override def cancel(cause: Throwable): Unit = ()
-    override def isClosed: Boolean = !hasNext
-  }
-
-  private final class InflightRangeSource[T](range: immutable.Range) extends InflightSource[T] {
-    private val isEmptyRange = range.isEmpty
-    private val rangeLast = if (isEmptyRange) 0 else range.last
-    private val rangeStep = range.step
-    private var nextElement = range.start
-    private var closed = isEmptyRange
-
-    override def hasNext: Boolean = !closed
-    override def next(): T =
-      if (closed) throw new NoSuchElementException("next called after completion")
-      else {
-        val current = nextElement
-        if (current == rangeLast) closed = true
-        else nextElement = current + rangeStep
-        current.asInstanceOf[T]
-      }
-    override def tryPull(): Unit = ()
-    override def cancel(cause: Throwable): Unit = ()
-    override def isClosed: Boolean = closed
-  }
-
-  private final class InflightRepeatSource[T](elem: T) extends InflightSource[T] {
-    override def hasNext: Boolean = true
-    override def next(): T = elem
-    override def tryPull(): Unit = ()
-    override def cancel(cause: Throwable): Unit = ()
-    override def isClosed: Boolean = false
-  }
-
-  private final class InflightCompletedFutureSource[T](result: Try[T]) extends InflightSource[T] {
-    private var _hasNext = result.isSuccess
-    override def hasNext: Boolean = _hasNext
-    override def next(): T = {
-      if (_hasNext) {
-        _hasNext = false
-        result.get
-      } else throw new NoSuchElementException("next called after completion")
-    }
-    override def hasFailed: Boolean = result.isFailure
-    override def failure: Option[Throwable] = result.failed.toOption
-    override def tryPull(): Unit = ()
-    override def cancel(cause: Throwable): Unit = ()
-    override def isClosed: Boolean = true
-  }
-
-  private final class InflightPendingFutureSource[T](cb: InflightSource[T] => Unit)
-      extends InflightSource[T]
-      with (Try[T] => Unit) {
-    private var result: Try[T] = MapAsync.NotYetThere
-    private var consumed = false
-    override def apply(result: Try[T]): Unit = {
-      this.result = result
-      cb(this)
-    }
-    override def hasNext: Boolean = (result ne MapAsync.NotYetThere) && !consumed && result.isSuccess
-    override def next(): T = {
-      if (!consumed) {
-        consumed = true
-        result.get
-      } else throw new NoSuchElementException("next called after completion")
-    }
-    override def hasFailed: Boolean = (result ne MapAsync.NotYetThere) && result.isFailure
-    override def failure: Option[Throwable] = if (result eq MapAsync.NotYetThere) None else result.failed.toOption
-    override def tryPull(): Unit = ()
-    override def cancel(cause: Throwable): Unit = ()
-    override def isClosed: Boolean = consumed || hasFailed
-  }
-}
 
 /**
  * INTERNAL API
@@ -138,7 +46,6 @@ private[pekko] final class FlattenConcat[T, M](parallelism: Int)
   override val shape: FlowShape[Graph[SourceShape[T], M], T] = FlowShape(in, out)
   override def createLogic(enclosingAttributes: Attributes) = {
     object FlattenConcatLogic extends GraphStageLogic(shape) with InHandler with OutHandler {
-      import FlattenConcat._
       // InflightSource[T] or SingleSource[T]
       // AnyRef here to avoid lift the SingleSource[T] to InflightSource[T]
       private var queue: BufferImpl[AnyRef] = _
