+++ stream/src/main/scala/org/apache/pekko/stream/impl/fusing/StreamOfStreams.scala
@@ -39,90 +39,10 @@ import pekko.stream.impl.Stages.DefaultAttributes
 import pekko.stream.impl.SubscriptionTimeoutException
 import pekko.stream.impl.TraversalBuilder
 import pekko.stream.impl.fusing.GraphStages.{ FutureSource, RepeatSource, SingleSource }
+import pekko.stream.impl.fusing.InflightSources._
 import pekko.stream.scaladsl._
 import pekko.stream.stage._
 import pekko.util.OptionVal
-
-/**
- * INTERNAL API
- */
-@InternalApi
-private[pekko] object FlattenMerge {
-  // Lightweight in-memory representation of a value-presented Source so the
-  // breadth slot can be occupied without paying for substream materialization.
-  private sealed abstract class InflightSource[T] {
-    def hasNext: Boolean
-    def next(): T
-    def isClosed: Boolean
-    def cancel(cause: Throwable): Unit = ()
-    def hasFailed: Boolean = failure.isDefined
-    def failure: Option[Throwable] = None
-  }
-
-  private final class InflightIteratorSource[T](iterator: Iterator[T]) extends InflightSource[T] {
-    override def hasNext: Boolean = iterator.hasNext
-    override def next(): T = iterator.next()
-    override def isClosed: Boolean = !hasNext
-  }
-
-  private final class InflightRangeSource[T](range: immutable.Range) extends InflightSource[T] {
-    private val isEmptyRange = range.isEmpty
-    private val rangeLast = if (isEmptyRange) 0 else range.last
-    private val rangeStep = range.step
-    private var nextElement = range.start
-    private var closed = isEmptyRange
-
-    override def hasNext: Boolean = !closed
-    override def next(): T =
-      if (closed) throw new NoSuchElementException("next called after completion")
-      else {
-        val current = nextElement
-        if (current == rangeLast) closed = true
-        else nextElement = current + rangeStep
-        current.asInstanceOf[T]
-      }
-    override def isClosed: Boolean = closed
-  }
-
-  private final class InflightRepeatSource[T](elem: T) extends InflightSource[T] {
-    override def hasNext: Boolean = true
-    override def next(): T = elem
-    override def isClosed: Boolean = false
-  }
-
-  private final class InflightCompletedFutureSource[T](result: Try[T]) extends InflightSource[T] {
-    private var _hasNext = result.isSuccess
-    override def hasNext: Boolean = _hasNext
-    override def next(): T =
-      if (_hasNext) {
-        _hasNext = false
-        result.get
-      } else throw new NoSuchElementException("next called after completion")
-    override def hasFailed: Boolean = result.isFailure
-    override def failure: Option[Throwable] = result.failed.toOption
-    override def isClosed: Boolean = !_hasNext
-  }
-
-  private final class InflightPendingFutureSource[T](cb: InflightSource[T] => Unit)
-      extends InflightSource[T]
-      with (Try[T] => Unit) {
-    private var result: Try[T] = MapAsync.NotYetThere
-    private var consumed = false
-    override def apply(result: Try[T]): Unit = {
-      this.result = result
-      cb(this)
-    }
-    override def hasNext: Boolean = (result ne MapAsync.NotYetThere) && !consumed && result.isSuccess
-    override def next(): T =
-      if (!consumed) {
-        consumed = true
-        result.get
-      } else throw new NoSuchElementException("next called after completion")
-    override def hasFailed: Boolean = (result ne MapAsync.NotYetThere) && result.isFailure
-    override def failure: Option[Throwable] = if (result eq MapAsync.NotYetThere) None else result.failed.toOption
-    override def isClosed: Boolean = consumed || hasFailed
-  }
-}
 
 /**
  * INTERNAL API
@@ -138,7 +58,6 @@ private[pekko] object FlattenMerge {
 
   override def createLogic(enclosingAttributes: Attributes) =
     new GraphStageLogic(shape) with OutHandler with InHandler {
-      import FlattenMerge._
       var sources = Set.empty[SubSinkInlet[T]]
       var pendingSingleSources = 0
       var pendingInflightSources = 0
